Ultimo commit do trabalho!

+++ projeto_pratica_api/models/MM_Lugares.cs
@@ -5,8 +5,8 @@ namespace projeto_pratica_api.models
         public int Id {get; set;}
         public string nomeLugar{get; set;}
         public string nomeCidade{get; set;}
-        public int coordX {get; set;}
-        public int coordY {get; set;}
+        public double coordX {get; set;}
+        public double coordY {get; set;}
         public int IdPais {get; set;}
     }
 }
